/vin endpoint on NFTs (#391)
* Updating vin endpoint

* Fix some stuff up

* Swagger

* Fix test

* Fix body Swagger for VIN update

---------
